Driver location stored after shift hours #7000

+++ include/library/Crunchbutton/Admin.php
@@ -467,8 +467,7 @@ class Crunchbutton_Admin extends Cana_Table_Trackchange {
 			} else {
 				$hours = 0;
 			}
-
-			return intval( $hours );
+			return floatval( $hours );
 		}
 		return false;
 	}
